Create Rubix cube
Standard 3x3 Rubik’s cube is visualized and builds. The user can perform all the clockwise and anticlockwise transformation on the virtual Rubik’s cube that he/she can perform on the real one. The user can also rotate the cube to see other faces with the help of mouse motion.