Create Library Management System in Python
The main objective of the Library Management system project is discipline of the planning, organizing and managing the library tasks. Our project aims at making the task of library easy. Library Management is entering the records of new book and retrieving the details of book available in the library. We can issue book to the library member and maintain their records and can also checks how many book are issued and stock available in the library. In the project we can maintain the late fine of library member who return the issued book after the due date. This Library management system project in Python,Oracle Datasbase Server,ORM(i.e. sqlAlchemy).